[GUI] Combine the overlay classes into a single header (As they are very small anyways and they all work together, it makes sense to keep them together)
