+++ src/client/lua/gui_api.cpp
@@ -3,7 +3,6 @@
 #include <common/scripting/script_engine.h>
 
 #include "../gui/overlay.h"
-#include "../gui/overlay_factory.h"
 #include "../renderer/gui_renderer.h"
 
 namespace {
